chore: move HandleProtocolAdapter to its own module
